+++ sentry/src/main/java/io/sentry/protocol/Device.java
@@ -99,7 +99,7 @@ public final class Device implements IUnknownPropertiesConsumer, Cloneable {
   private Float batteryTemperature;
 
   @SuppressWarnings("unused")
-  private Map<String, Object> unknown;
+  private Map<String, @NotNull Object> unknown;
 
   public String getName() {
     return name;
@@ -350,7 +350,7 @@ public final class Device implements IUnknownPropertiesConsumer, Cloneable {
 
   @ApiStatus.Internal
   @Override
-  public void acceptUnknownProperties(Map<String, Object> unknown) {
+  public void acceptUnknownProperties(Map<String, @NotNull Object> unknown) {
     this.unknown = new ConcurrentHashMap<>(unknown);
   }